Change world constructor to take block amount

+++ src/chunk/world.hpp
@@ -6,8 +6,8 @@
 namespace chunk {
 class World {
 public:
-  World(const ::core::vulkan::Context &context, const size_t width,
-        const size_t depth);
+  World(const ::core::vulkan::Context &context, const int width,
+        const int depth);
 
   void place_block(const glm::ivec3 &position,
                    const BlockArray::BlockType block);
